settled on the temporary bridge api
