+++ dimos/protocol/pubsub/test_pattern_sub.py
@@ -25,8 +25,8 @@ from typing import Any, Generic, TypeVar
 import pytest
 
 from dimos.msgs.geometry_msgs import Pose, Quaternion, Vector3
-from dimos.protocol.pubsub.impl.lcmpubsub import LCM, Glob, LCMPubSubBase, Topic
-from dimos.protocol.pubsub.spec import AllPubSub, PubSub
+from dimos.protocol.pubsub.impl.lcmpubsub import LCM, LCMPubSubBase, Topic
+from dimos.protocol.pubsub.spec import AllPubSub, Glob, PubSub
 
 TopicT = TypeVar("TopicT")
 MsgT = TypeVar("MsgT")
